Avoid stat calls in _fn_version
This call is completely unnecessary. If it's a plugin, then if it
doesn't exist it won't be in modInfos, so modInfos.getVersion will
handle it correctly. If it's a binary, then get_file_version will handle
it correctly (pywin32 will throw an error, which that method will
catch).

Avoid stat calls in _fn_product_version

Same as above.