WIP trying to call wx.Locale once:
This seems to work but leaves a lot to the imagination. Anyway since
python itself has a slew of issues on its tracker for locale handling
that's all we can do for now. At least we don't set and reset the Locale
(an expensive operation, and meant to be done once). See also:

https://github.com/wxWidgets/Phoenix/issues/769#issuecomment-1063235383

for a reproducer of the bug(s) - can't be reliably reproduced which
probably shows there is a race somewhere deep inside the wx C code.

Restore the InitLocale but only restore LC_TYPE to C locale

See:
https://bugs.python.org/issue38805#msg373896

Closes #610 - for now