Initialise ColorChecks once:
Since we have a wx.App and the only dir we need here is dirs['images']
which is set in init_mopy_dirs we can finally drop this hack - this was
triggered by the huge bugdumps where the same images were read from disc
once per tab. We still need to address #366 but getting there.
Mopy/bash/basher/mod_links.py: regression introduced in
28c07f704e8585b863bc1071a440e3fc13c9bb7d

Ok third's the charm - this emulate_startup needs rethinking, as well
as setting the working dir - probably the right way is to set the dirs
correctly in init_mopy_dirs and drop __get_translations_dir but this
will do for now.