setGhost drop ghost state calculation:
The difficulty is tracing the _lo_cache decorator the rest of the uses
seem safe. Initially I had added the itsa_ghost parameter, but seeing
that we always passed the ghost state of the info we were called on,
and that _refresh_ghost_state would just set the isGhost to the value
we passed, therefore a no-op I went on and dropped the itsa_ghost:

@@ -665,3 +665,3 @@ def abs_path(self):

-    def setGhost(self, isGhost, *, itsa_ghost=None):
+    def setGhost(self, isGhost):
         """Sets file to/from ghost mode. Returns ghost status at end."""
@@ -672,4 +672,2 @@ def setGhost(self, isGhost, *, itsa_ghost=None):
         # Both should never exist simultaneously, Bash will warn in BashBugDump
-        # TODO(ut): we always pass itsa_ghost - drop the call below!
-        self._refresh_ghost_state(itsa_ghost=itsa_ghost)
         # Current status == what we want it?

Mopy/bash/basher/mod_links.py: the itsa_ghost param was added on dev in
4cb1cbf78f815913b70bb32c91a8c3917f2482a3 mistakenly - yey for
backporting.

Mopy/bash/basher/__init__.py: stray import from parent