ModInfo.in_master_block:
This was essentially static but this is WIP - for one I am not sure the
master_block heuristics is correct - that  was a long time ago. Maybe we
need a ModInfo attribute like _is_esl. Triggered by the coupling of
LootParser with load_order -> bush, but since then loot_conditions
started importing bush while load_order stopped.