[!]Drop table:
A big step towards finishing off #336. In fact this was an old branch I
unearthed while working on new_info - I had enough going through the
TableFileInfos overrides of key API methods, which were only there
to update the table. Turns out table had some very deep roots - in
particular the Info instance was linked to its table properties by its
fn_key, so carrying the table properties to an updated info with the
same filename was done implicitly. This was not always intended, for
instance resulted in needing to force the crc calculation as the crc
cache was carried along unwillingly. Now the updates of the properties
must be done explicitly, which is the right thing to do, but still wip -
finalized in new_info refactoring a few merges ahead. For now I
introduced a half baked API (copy_persistent_attributes) - half baked
partly cause I am not sure yet about what we should do in a case by
case basis for some of ModInfo attributes and also cause some of these
attributes like autoGhosting must be already set during refresh -
hence [!] and let's make haste and merge the new_info changes that
will make it possible to set_table_prop in init time. I realized how
deep the roots of table were starting from the report on BP config not
being read:

https://discord.com/channels/537706885965676554/537710082755133460/1232065762676572272

What was happening was _update_info_sets reading the config from info
while it was set for patchInfo - same name so previous code would read
it from the table while now we need to copy_persistent_attrs(patchInfo).
I bet that bp_split_parent would not be set correctly either now that
table is gone - however with table it was almost accidental that it
worked.

There are other places we should be setting the table attributes on the
new_info - those are few and we must decide on a case by case basis.
It's not trivial to decide which of the _key_attr's keys need to be
passed to refresh, but in most cases (excluding rename which does not
call new_info just edits existing) some of the table attributes should
*not* be passed. For sure ['crc', 'crc_mtime', 'crc_size'] most likely
['ignoreDirty', 'mergeInfo'] and 'bp_split_parent' should be ok - but
we *should* include 'autoBashTags', 'bashTags', as far as I understand -
do we want to rescan for Bash tags when reverting a bakup ?[!]. Note
that the crc info being carried along in _revert_backup (and when
using new_info in similar code like snapshots) looks like a bug - in
the BP was explicitly addressed - now we start with 'crc' None and we
don't need calling calculate_crc explicitly nor pass crc_changed=True
in those cases(yak, yey). The API limits were table - will be
completely fixed once __init__ can set table properties, hopefully
till then it will be clear what should be in _ignore_on_revert. Note:

- delete_refresh - nothing to do - yey!
- same for rename operation (as the parent method edits the info
inplace) - more on renames soon
- add_info has the problems of new_info - more on that later
- gone the hackiest use of get_store(), which is itself a hack (the
infos should not need to know their parent - used for backup/hide and
notifying BAIN, beta stuff).
- I set tracks_ownership = False for SaveInfos - should not be needed
as BAIN is not configured to install ess files but it's correct - anyway
the interplay between BAIN and _AfileInfos is post #336
- ignore the fixme in save_pickle see next commit

Last use of 'mtime' col was in b97759ae6a3f3ea88dd1aacdf7e8610a00f0002f
from what I can gather