TES5, FO4: Disable save header editing
We can't actually do this safely, because we have to handle ESL and
regular masters separately. Right now we don't, which will just horribly
break the master list if you do try doing it. A future commit lays some
groundwork for this by splitting the masters list for such save headers
in two, but it's not enough yet.