Allow ESM-flagged plugins to be ESL-capable
A remnant from back when we didn't know how the ESL flag properly works.
An ESM with the ESL flag is of course completely fine - that's how all
of Bethesda's creation club content works, after all.

I also removed the discard() call - we just clear()ed self.mergeable, so
the only way this could do anything is if we had the same mod in
ModInfos twice - which is impossible.