converters: Fix various issues with BCFs
Deprint if we move a corrupt BCF

Otherwise it's pretty much impossible to figure out why a BCF was deemed
'corrupt'.

Avoid messing with encodings in the translator

We just need to swap the literal bytestrings 'bosh' and 'bolt' for
'bash.bosh' and 'bash.bolt', respectively. Was causing a traceback
beforehand.

Fix BCFs getting overridden by output archive

The path we were passing to 7z was wrong and would replace the BCF
itself with the newly compressed output archive. Also took the
opportunity to bin the pointless dest_dir parameter of compress7z - 7z's
'a' command does not even take a '-o' switch, so that parameter was
wholly useless: https://sevenzip.osdn.jp/chm/cmdline/commands/add.htm