TES5: Port Tweak Names
Ports the last patcher to Skyrim (barring FormID Lists, but that one is
pretty broken and requires reworking - see #497). Eight years after #151
was opened, we can finally close it.

Sort: Ingestibles can be ported, but is pretty pointless because all but
three of the potions in Skyrim have no magic school set in their
effects, so they all get the 'U' code. On top of that, the UI already
splits food and potions into separate sections, so there's no need for
that part of the tweak either.

Most other tweaks ported easily, except for the initial challenge of
figuring out that I had to use the half-cost perks to determine the
'level' of each spell.

Mopy/bash/images/people/ifniavi48.png:
Update my own avatar

Haven't used that one for close to a year now.

Closes #151