Validate FOMOD XML schema when running FOMODs
This adds lxml as an optional dependency. If it's present, we use that
instead of Python's stdlib XML parser (the API is nearly 100%
compatible). We can then also validate FOMOD ModuleConfig.xml files
against the official schema, which is included in fomod_schema.py. Note
that I had to fix a bug in the schema itself before lxml accepted it
(lol) and that I included foseDependency directly, which was introduced
in 5.1.

A popup is shown if the validation fails that explains what happened and
contains a full log of errors that can be shared with the mod author. It
is also possible to continue with the installation regardless of the
errors.

Users can also disable verification globally through a new option in the
Installers' column menu/global menu.

Closes #603