Add Current Index column to Saves tab
As requested on AFKMods. Simply adds a column that shows the current
'true' in-game indices of the masters in the master list, as set on the
Mods tab. In other words, this just shows the value that would also be
shown on the Mods tab's Index column for the master in question.

Followup to 6af88dab37e8782a4980086c321f9ccd61c222f5.
Under #624