AIniFile - #336 WIP:
Note I unslotted AFile - need some more flexibility plus this is WIP.
Couldn't play with MI. I also unslotted Installer but that was the case
since forever - we are using the bosh overrides which never defined
slots, otherwise the fomod added attributes wouldn't compile (in
bytecode :P). Rest is of course IniFile refactoring - it-s a ListInfo,
my word, but not necessarily an AFile. This cleans the abstract dirt
from default tweaks (note I made is_default_tweak the class variable it
is). IniInfo hierarchy became sharper - note listErrors for instance
should not be in default_tweak (I hope). ListInfo got the get/set table
props to unify the use of this API for all the clients - that API being
refactored further down this WIP branch. It also gained fn_key - this
makes it beta :)

I thought of caching read_ini_content for DefaultTweak but it's really
not needed - used once on copying the ini which is a rare/do once
operation.