Show a BP config error when validate_values fails
If we somehow end up letting an invalid custom value through (this could
happen if a tweak used to accept a value in one version of WB, that
value then got saved into the BP and now a newer version of WB has the
tweak reject that value), simply fail on initializing the MultiTweaker.
The isinstance(str) check needed in py3 otherwise we get:

TypeError: '<' not supported between instances of 'str' and 'int'

Note in py2 the comparison would return False, which in this case was
the desired result.

The rest is py3 cleanup - typing, u'' removal and super() cleanup.

Under #618