Mel.load_bytes:
Finally some light at the end of the tunnel. Root of complications turns
out to be mixing the deserialization logic with assignment to record
logic. This is the final Mel API form - will be put to use in next
commit. However there is a performance hit - more function calls:

patcher_dialog.py  191 PatchExecute:
834049310 function calls (803111525 primitive calls) in 347.361 seconds

Once we get the API right this must be addressed (a metaclass that
inlines?)