MelFidList -> MelSimpleArray:
Turns out some of the MelFidList elements were used as lists of fids - plus
the idea of creating a bunch of MelObjects just to store a fid (or a number
for that matter) feels wrong. So some new machinery - this is by the way
the last piece of tools needed for a proper MelEl hierarchy. Turns out
most of the complications here are due to trying to have Mels serve as
both de/serializers and adapters (of the plugin file structure) to the
records representation. De-entangling is now possible at last.