Fix custom choice popup when types changed
If a type (like GmstTweak_Crime_PickpocketingChance, for example) has
changed type over time, older BP configs may contain custom values with
the old type (int in this case). It's not possible for a user to fix
such issues right now. This commit makes it possible by always looking
for the value types at the default values instead of the custom ones.