Mopy/bash/brec/basic_elements.py: SSS TTTime
MelStruct pack - only apply action if needed, it should be the exception
(actually we should strive to always wrap the action elements to their
action on load). Doesn't play well with the FID action (meant to
only be applied on load, as after loading the masters info is lost).

Rest is more fdust from future