Clean up Installers context menu further
This is pretty much as clean as it's going to get. Everything that's
archive-specific is now in an 'Archive..' sub-menu that only appears for
archives, everything that's project-specific is a 'Project..' sub-menu
that only appears for projects. The few remaining loose options that
hung around at the bottom of the context menu have been moved into the
'Package..' menu or into the archive/project menu if they were specific
to archives/projects (e.g. the 'Omod Info...' link, which is completely
project-specific and super rarely used).

Move Sync From Data to top level

Useful enough (e.g. for syncing after xEdit cleaning) that it should be
top level.