Handle wx.ICON_INFORMATION properly in askStyled
It previously showed a warning dialog instead of an info dialog. Not
sure why, vistaDialog absolutely supports showing an info icon.

Fixes e.g. the move/copy to profile commands on the Saves tab showing a
warning icon when they want to notify the user that their requested
action was completed successfully.

Also took this opportunity to swap out ICON_HAND for ICON_ERROR and
ICON_EXCLAMATION for ICON_WARNING, since the wx docs advise that the
former are compatibility(?) aliases for the latter.

Mopy/bash/bash.py:
Dump lxml version on boot

Forgot to add this when I added lxml as an optional dependency.