Handle inis consisting entirely of comments:
Mopy/bash/ini_files.py: see
https://discord.com/channels/537706885965676554/537710082755133460/969645213737959504
and posts below.

Initially I was just returning IniFile but that would not work - if the
ini was an OBSE one and was edited, we would not rerun the heuristics to
find out the type. Hence in this special case some more processing is in
order.

Infernio:

Slight rewrite to simplify and make it detect OBSE INIs correctly.

By default, behave exactly like before: strip off the whole comment and
only look for uncommented lines. Only if we find nothing that way and
have at least one comment do we strip off the leading ';' and spaces and
consider the commented out line as if it were uncommented.

Has false positives of course (same as in the original comment), but
since there are many more 'real' OBSE INI lines commented out, it finds
the correct type successfully.

Simplify by avoiding exceptions as control flow

Also drop the any() check, if there are no comments in the file we'll
know because count will be falsy again afterwards.

Co-authored-by: Infernio <infernio@icloud.com>