Add link to create a project from FOMOD output RRR
Can be useful to get started with porting an FOMOD-only mod to BAIN. For
projects, we can simply shell-copy the required files into a new
destination project and tell BAIN about the project. For archives, it's
more complicated since we first have to extract into a temporary
directory, then copy the files to their correct paths and finally tell
BAIN about the new project.

Closes # 593 <--- RRR