Fix regressions from parsers work
Not sure when this broke, but exporting anything (e.g. Stats) was
broken. First was a type error due to accessing a lambda via self rather
than self.__class__ (the former passes self due to being a bound
method). Then a TypeError due to trying to hash a list (not sure about
the fix, it's the values_key -> v one) and finally f-strings don't work
with Flags due to it not implementing __format__.

Mopy/bash/basher/mod_links.py:

Show error when exporting with missing masters

Rather than cryptically failing with a KeyError when we try to access
data for the missing masters.

parsers.py: Clean up newline handling a bit

Was causing extra newlines to be spat out for Factions and Relations
CSVs.