Allow New Project to create ModuleConfigs
Some more minor FOMOD support in BAIN. Also cleaned up the code and some
PEP8'ing. And as a bonus, made the blank wizard.txt a bit less blank by
including an example SelectPlugin statement in it.

Note I changed the behavior of creating a Blank.esp/Blank Masterless.esp
to allow both to be created for a single project. That made the logic
for the example wizard generation simpler and can't really hurt IMO.