TES3: Handle {CREA,NPC_}\FLAG correctly
This isn't really a full flags field - the first byte is, but the later
bits are weird combinations corresponding in an enum-like fashion to
blood types/textures defined in the INI, according to the UESP page.
Just treat it as 3 random bytes for now and restrict the flags to the
first byte.

Mopy/bash/game/skyrim/records.py:
TES5: DOOR\TNAM should be sorted