Typing motivated fixups:
fixup: method signature for `env.linux.TaskDialog`

Needs the same signature across implementations, otherwise calling
using keyword arguments won't work consistently across platforms.

private `SaveFileHeader.load_masters`

`load_masters` leaves `.masters` as a `list[bytes]`, but the public
API is for it to be a `list[FName]`.  This is only every called
internally so far, and if you need to call it yourself you'd have to
know to call `_decode_masters` as well.  Better to private it.

Under #634