[!}Fix for never passing a crc_installer to build:
Note Mopy/bash/basher/installer_links.py never passed a crc - we could
inline build

Note I replaced the "cext in" with "endswith" - it is 5 times faster
and chops off the GPaths - but the whole subArchive thing was short
circuited (only place used was this build method) so needs testing
(law of conservation of TODOs - !).

-seems subArchives return value is only used in a Path.join so no need
to GPath.

Partly reconstruct bcfPath_sizeCrcDate to avoid a full refresh on boot:

This really avoids creating the Progress objects if crc did not change
(we still need to crc/stat to update the cache).

Class factories:

- don't re stat in addConverter
- avoid recalculating the crc