Skip userlists if parsing them fails
Userlists can end up with all kinds of wild formats that will then also
cause all kinds of wild errors. Sometimes YAML parsing fails, sometimes
it doesn't (but does cause errors when we try to interpret the structure
at runtime, e.g. an IndexError or a KeyError). Userlists aren't that
important compared to the main masterlist, so skip these and complain
about them in the BashBugDump instead of crashing.