Forward BOOKs in Cobl Catalogs
Otherwise we may discard valid changes made to the books (e.g. to change
their covers). I also asked around on Discord and it seems that removing
the script was probably an oversight(?) so I did not add any code to
keep the existing behavior there.

Nice side effect is that the code is much simpler and shorter now. Not
sure the return None/is None stuff is needed, but I didn't want it to
crash if someone's using a weird version of Cobl that lacks the BOOKs in
question (there must have been a reason we were making new records after
all).