Censor the user's home directory when deprinting
Knowing the user's home directory was useful back when we had tons of
unicode problems with it in py2, but we're on py3 now which does a
*much* better job. At this point, deprinting the user's home directory
only acts as a vector for people to unknowingly doxx themselves by
revealing their real names inside publicly accessible log files (anyone
can download and view BashBugDumps posted to
Discord/AFKMods/Reddit/whatever).

So censor them by replacing the the username with '*****'. Should work
on both Windows and Linux, but I haven't tested it on Linux yet.

Mopy/bash/loot_parser.py:
Make the userlist except match the masterlist

Just a minor improvement to the message and handling catch_errors
correctly (useful for future unit tests).