Performance - drop callable check for MelStruct actions:
action is either zero or a callable that would evaluate to True - does
chop off some 5 secs/340 secs in my test BP.

Infernio:

Default actions to None

Otherwise we break Morrowind on boot, since [Auto]FixedStrings evaluate
to False in an if if they're empty. Also took the opportunity to support
specifying an action for float attrs. Don't think we use that yet, but
it seemed weird to me to silently overwrite an action there and would
probably lead to a fun bug chase if we ever do end up needing it.

Co-authored-by: Infernio <infernio@icloud.com>