Remove redundant _wx_widget_type
We can use __annotations__ to retrieve the type of _native_widget and
use that. Saves a bunch of redundant definitions and removes a potential
source of errors (a mismatch between _native_widget and
_wx_widget_type). Note that I had to add an override mechanism for four
classes, wx.html2.WebView, wx.TreeCtrl, wx.RadioButton, wx.CheckBox. For
WebView the reason is clear (we have to call WebView.New, not
WebView.__init__), but for the other three __annotations__ contains only
a string and I'm not sure why. I thought it might be the annotation on
the parent for RadioButton and Checkbox, but I tried removing or
simplifying it and the problem persisted :shrug: