Fix 'Revert to Saved' not working correctly
This was a hell of a mini-rabbit hole. Fixing it for 'Revert to Saved'
broke it for 'Revert to Default'. I'm still not convinced this is the
proper fix - this whole 'loading config to both restore and revert'
thing is quite hacky and should probably be two different methods
instead.