Drop debug mode, tee deprint to dump and stdout
The BashBugDump is very helpful, but has one huge downside: activating
it gets rid of the 'native' error reporting mechanism (most notably
wxPython's popup window), which means you have to remember to check the
BashBugDump every time you close WB.

We enabled the BashBugDump by default for standalone builds a while ago,
but by doing so, lost wxPython's error reporting. This makes it harder
for users to notice when something is broken.

This commit gets rid of debug mode entirely and always enables debug
printing. It also introduces bolt.Tee, which is used to copy the output
to both the native stdout (which means you can now see BashBugDump
output in the PyCharm console) and the BashBugDump.

Closes #638