Drop FormID mismatch checks in CELLs and DIALs
CELL and DIAL FormIDs don't need to be compared - we index the Mob*
classes based on those FormIDs, so they're guaranteed to be equal. Those
checks are just wasted performance.

PGRD and LAND are more interesting. The game treats any occurence of
these as a CELL child the same way, whether they are new records or
overrides. So it is legitimate to have conflicting FormIDs here and we
can't check it.

This means that the checks in MobCell and MobDial have to go. The checks
in MobWorld probably do as well, but I can't say that for sure yet, so
leaving them in - let's see if it blows up at some point.

Improve display of ModFidMismatchError

Should show the FormID as hex via strFid.