Update all header (i.e. TES4) definitions
Turns out OFST and DELE were missing from some of them, TES4 had an
extra DATA MelNull in there for no reason and the FO4 one did not have
TNAM, INTV and INCC correctly decoded.

Required adding the prelude parameter to MelSimpleArray, since I don't
want FO4 to create a ton of MelObjects for the transient types in the
header at boot. We probably should use MelSimpleArray more often
(basically whenever we don't *need* a MelObject for the array, it should
be a MelSimpleArray).