FO4: Implement ACTI
I chose not to refactor, update and use our VMAD implementation yet and
instead discard all VMAD for now. The reason is that I'm not sure how
badly doin that would conflict with the ongoing FormID work. Plus it
would have bloated this commit beyond recognition. It's already big
enough as it contains refactoring of some shared subrecords between the
games (yay for reducing occurences of those string constants!).