Add a (debug) way to tell the real file position
ModReader now wraps the input stream in a BytesIO object for
performance, but this resets the tell() position back to 0. Keep track
of a debug offset and add a tell_debug() method to be used when
debugging so that you don't have to manually keep track of these offsets
when investigating the binary file itself in a hex editor.