TES5: Implement NAVM
Note that this can't be merged no matter what since it's intimately
linked to NAVI. The same is true for FO3/FNV and FO4 (we already don't
allow merging it in FO3/FNV, so everything is fine there). Also still
can't be loaded since some TES4-only code prevents us loading CELL
children that weren't around in TES4, so this is really only an internal
commit.

Slightly cleaned up the init() methods while I'm at it.