bolt.fast_cached_property
functools.cached_property features thread safety via locking for some
reason. The Python core devs have acknowledged that this was bad design
(which costs us measurable performance - about 10s on my Skyrim BP which
takes 5min, so ~3%), see CPython bug 87634. The activity on that issue
seems to be going in the direction of adding a kwarg like nolock=True to
avoid this.

We can gain some more performance by avoiding some other features of
cached_property that we don't need.