common_subrecords: Sort and update for py3
That's it, I've had enough. Every time I deduplicate a subrecord and
move it here, I'm stuck wondering where to put it. This file is such a
half-sorted mess, and I'm going clean it up now.

As this left only ~40 lines in the file that needed py3 followup, I
decided to just do that with the whole file (u'' prefix removal and
changing super(bla) to super()).

Plus some minor refactoring in MelRelations and MelWeatherTypes.

Under #618