FO4: Implement BOOK
Note I dropped the unused 'modb' slot for all MreBooks. Been there since
they were added as far as I can tell, but there's no reason for it to be
there - that's not even attribute name we use (outside of MelModels, but
those sit in their own group and so don't use these main slots).