FO4: Implement BPTD
MelUnorderedGroups already pulling its weight by getting rid of that one
hack in the FO3/FNV record definitions. There are probably other
MelGroups that should really be MelUnorderedGroups as well, but that
would require tons of investigation and testing via the CK to find which
subrecords are really required (xEdit's definitions are often not
correct there).

And with that, all B* record types are done.