Fix 543 regression in BCFs
Would pickle FNames into the BCF, which obviously get saved as Paths.
This would blow up when trying to apply the BCF. Fixed it so that BCFs
created this way can still be applied and fixed the code that was
writing FNames into the BCFs in the first place so that we once again
only write strings for backwards compatibility.