Implement 'fused attributes' for APreserver
The 'Level' attribute and the 'PC Level Offset/Mult' flag are
inextricably linked: the flag changes how the attribute is interpreted,
so carrying the two of them forward separately results in completely
wrong output records (e.g. a way too low PC level multiplier).

Turns out there are still some more features to add to our lovely
preserver base class. Fused attributes are used by simply replacing a
regular attribute (i.e. a string in the final tuple) with a tuple of
strings. attrgetter_cache already understands this (since I needed that
back when I implemented MelSorted's sort_by_attrs), so this is really
just down to making the buildPatch phase understand it as well. Also
took the opportunity to implement a TODO from # 312 <--- RRR - moving
the attrgetter_cache access out of _init_data_loop's for loop for some
perf gains (and moved the whole 'FormID attrs' check out of the loop by
duplicating the loop instead - that way we avoid the checks entirely
while inside this extremely hot loop).

Benchmarking shows the perf gains are minor (within margin of error),
but it must be faster since we're doing less work in one of the hottest
loops in the entire codebase.

Under #312