Drop bolt.{un,}pack_string
The struct thing we were doing in there was wholly pointless - ins.read
already returns a perfectly good bytestring, struct-packing it right
back into a bytestring is just thrown away performance. And after
removing that, this is now literally just a fancy lambda, so move it to
its only usage in save_headers and refactor for some perf gains too.
Py3ified lines I touched and ones in the immediate vicinity.