FO4: Implement EFSH
What a terrible record. Reformatted and PEP8'd all of it, but it's still
awful. In FO4, Beth decided to clean it up - but they kept the old
format around as well (with minor changes, so I can't just deduplicate
it...).

This record may also need to skip form version upgrade - or it may work
fine. We'll have to test later on.