Add Ctrl+S shortcut and Save Data link
Frequent saving is good practice, of course, so let's support that with
a more convenient way to do it (for reference, the current way to do it
is via Settings > Global Settings > Backup > Save Data or by exiting WB
- not ideal).

Left the link undocumented to not make the mess in there even worse - as
already mentioned in the redesign commits, we'll want to do all our GUI
redesign and changes, then redo screenshots and rewrite docs in one go,
otherwise we'll get swamped in duplicate work.

Under #645 for adding Global Settings to column menus.