Use FormId.from_tuple for Cobl fids
Definitely needed for the 'in dict' check, that won't work if the types
differ. Might be needed for the others too, but I honestly just think
it's a good idea to have them always be fids, even if just for docs
purposes.

Mopy/bash/basher/dialogs.py:
Add WB icon to New Plugin dialog