Fix 637 regression when sorting FormIDs
First off, this is a hack. The right solution is to stop sorting long
FormIDs entirely, but that will require us to rewrite _AMerger and
absorb the other mergers first (see #497).

That said, before the FormID handling rewrite, sorting FormIDs would
perform an alphanumeric sort while they were in long form, but a
master-order-dependent numeric sort while they were in short form. One
of these is reasonable (hint: it's the second), the other one isn't.
It's used nonetheless in _AMerger and _AListsMerger, where it's a key
component of the logic.

So for now, repair this regression by emulating the previous behavior.
First try short-mapping the FormIDs and comparing that. If that doesn't
work (because we're not in a write context), fall back on the current
alphanumeric comparison.