bain: Fix global start skips not using separators
If you had a plugin called 'ss_foo.esp' and enabled Skip Screenshots, it
would get skipped because the screenshots skip did not include a
trailing path separator. A couple start skips did, but most of them did
not. Fixed it by centralizing the place where the separator gets added
and removing it from the few random skips that did have it.