bain: Supports paths longer than 255 chars
Needs a registry value set, which we can do with the installer. Not
supported by our minimum supported Windows version (Windows 8) AFAIK,
but the registry key will just be ignored in that case. Plus, who
actually uses Windows 8/8.1?

Drops a nasty hack in BAIN and fixes a refresh loop since another part
of the code didn't have this hack applied to it.

Co-authored-by: Infernio <infernio@icloud.com>