Sort subrecords after loading as well
Just ran into a situation that showed this is necessary for proper
conflict resolution in the BP. Specifically, it was creating bogus
keyword (KWDA/KSIZ) overrides because a plugin had a different order of
KWDA subrecords than the master (Skyrim.esm), but the actual contents of
the KWDA subrecords was identical. I've been suspecting that this is
necessary for some time now (see also the regions TODO that was dealt
with here), so it's at least nice to get confirmation.

This does of course cause issues since we'll be sorting FormIDs there -
we really need to rethink this, clearly sorting FormIDs is necessary and
commonly used. The current 'alphanumeric handling during loading,
short-mapped integer sorting during dumping' works fine for now (and I'm
struggling to think of a proper solution beyond it), but there surely
*has* to be a better one.

Added needs_sorting, filtering out subrecords that way improves
performance by ~2x - now just 0.2s spent sorting all subrecords in
vanilla FO4, including all DLCs.