[!]Removing defaults from MelOptStruct:
Are any of those needed? Anyway these are the only explicit defaults set
here (apart the None as in "None here are on purpose") - tagging a ! in
case there was a reason 'obme_unused' needs not  be dumped if it is
null1 * reserved_byte_count (duh :P) - the rest seems pretty random.
The few remaining explicit defaults are None which will be the default
default (sic) soon. MelOptStruct works as before, not dumping if equal
with the MelStruct defaults (0). On flipping those to None all Structs
will become optional so we are about to retire MelOptStruct.