Move patcher constants to the GameInfo subclasses:
I initially tried to move condition_function_data to the records where
it belongs but the current model does not allow this - actually, it does
not allow to set the game more than once safely - see:

https://github.com/wrye-bash/wrye-bash/issues/400#issuecomment-1269682542

However because we need to at least test that record definitions load,
and because this was not allowed due to the dynamic import I here retire
this venerable hack (and add the tests in child commit).

Note:
- I zap newlines
- got rid of the def _fid/ del _fid
- removed u' in ported code - that was not a good idea, blurs the diff
(git should detect a "rename" here) and I accidentally renamed stuff
like texture_count_u  -> texture_count_ duh TTT
- manually formatted (needs some more)

A few patcher packages gone - the need to define an empty package was
an old smell.