Check legacy paths only in Oblivion
These are ancient and were there in our initial commit. Can't remove
them of course, since it could break perfectly functional installations
of WB for Oblivion. But we also don't have to check these for any game
except Oblivion - that includes Nehrim and Oblivion WS.

The impetus for this was that the Epic Games Store version of Skyrim SE
ships an 'Installers' folder with DirectX redistributables in it. WB saw
this and happily went 'aha, a legacy setup', when that's of course
nonsense.

Under #648