Change getActiveRecords to also return the record fid:
Note that the fid is not the key MobObjects uses. However in most
(all?) cases the returned "rid" (as in record id - reserved!) is used as
a fid (rid.mod_fn for instance). In the few cases that records are keyed
by eid this would blow - the fact it does not means that those few
record types are not accessed by the code but the model is very
fragile. We should:

- return the used (in id_records) record id or subclass to handle
those - aka iterate over id_records.items(). Even more so as 'records'
will be removed and id_records will be the core MobObjects data
structure
- revisit uses (see also "Bye MobObjects.records" commit) - we sometimes
use  getActiveReccords sometimes directly iterate over all records -
why (edit: probably on scanModFile we only keep non-ignored, but then
so later in the BP we don't need to filter again?)

Few (ancient) uses where record.fid was not accessed (in faces.py for
instance, I used _rid there)
Some 'rid' renames - \b([a-z_]+)?rid\b catches some `grid`s too

And last minute fixup:

@@ -227,3 +227,3 @@ def getActiveRecords(self):
         """Returns non-ignored records - XXX what about isKeyedByEid?"""
-        return [(r.fid, r) for r in self.iter_records() if
+        return [(r.fid, r) for r in self.id_records.values() if
                 not r.flags1.ignored]

MobObjects is inherited by MobDial - this is a mistake - MobDial has-a
MobObjects - this lead to several subtle bugs but it's tricky to uproot
without a complete rewrite (in the works).